// src/app/types/enums.ts
export enum Theme {
  DARK = 'dark',
  LIGHT = 'light',
}

export enum Skills {
  TYPESCRIPT = 'TypeScript',
  JAVASCRIPT = 'JavaScript',
  REACT = 'React.js',
  NEXT = 'Next.js',
  EXPRESS = 'Express.js',
  SVELTE = 'Svelte',
  ASPNET = 'ASP.NET Core',
  DOCKER = 'Docker',
  FIREBASE = 'Firebase',
  GIT = 'Git',
  POSTGRESQL = 'PostgreSQL',
  MONGODB = 'MongoDB',
  AWS = 'AWS',
  C = 'C',
}